Python/requirements.txt
azure-ai-vision-imageanalysis==1.0.0b3
azure-cognitiveservices-speech==1.36.0
azure-core==1.30.1
azure-identity==1.15.0
azure-keyvault-secrets==4.8.0
azure-storage-blob==12.22.0
Flask==3.0.0
Flask-SocketIO==5.3.6
Jinja2==3.1.2
openai==1.3.2
python-dotenv==1.0.0
pywin32==306
Werkzeug==3.0.1

